+++ example/.vuepress/config.js
@@ -1,5 +1,5 @@
+// .vuepress/config.js
 module.exports = {
-  theme: "vuepress-theme-minos",
   title: "Minos",
   description: "VuePress theme - Minos",
   head: [
@@ -53,5 +53,10 @@ module.exports = {
         link: "https://kencc.github.io/rss.xml",
       },
     ],
+    share: {
+      type: "sharethis",
+      installURL:
+        "https://platform-api.sharethis.com/js/sharethis.js#property=5e7ab9fe192d8a0012be9a58&product=inline-share-buttons",
+    },
   },
 };