+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# [1.0.4](https://github.com/kencc/vuepress-theme-minos/commit/8fdcd3fbf5c4d7d5c88ad9a43c432abf54a5b9c8) (2020-06-04)
+
+Feature Change
+
+- Restructure mSearchbox component in order to remove jQuery & Font Awesome dependency
+- Remove jQuery CDN dependency
+- Remove Font Awesome CDN dependency
+- Change to Router-Link for page switch
+- add initialize for shareThis plugin
+- Move all theme options from ./vuepress/theme/index.js to ./vuepress/config.js
+- Documentation update for About page
+
 # [1.0.3](https://github.com/kencc/vuepress-theme-minos/commit/92856a6bdb5e669d8b3ce9243ec3e39d5c7044d0) (2020-05-28)
 
 update documentations